Made the Feed Prettier, Split the exhibit into a group rather than just a card, in case of more info needing to be shown
