+++ src/components/feed/ExhibitGroup.js
@@ -1,0 +1,34 @@
+import React from 'react'
+import Button from '@material-ui/core/Button'
+import {Chip} from '@material-ui/core'
+import Card from '@material-ui/core/Card'
+import CardActions from '@material-ui/core/CardActions'
+import CardContent from '@material-ui/core/CardContent'
+import CardMedia from '@material-ui/core/CardMedia'
+import IconButton from '@material-ui/core/IconButton'
+import FavoriteIcon from '@material-ui/icons/Favorite'
+import Typography from '@material-ui/core/Typography'
+import ExhibitCard from './ExhibitCard'
+import '../../css/Exhibit.css'
+
+const ExhibitGroup = props => {
+  const { picture } = props
+  const { title } = props
+  const { summary } = props
+  const { tags } = props
+
+  return (
+    <div>
+      <div className='exhibit-center'>
+        <ExhibitCard
+        title={title}
+        picture={picture}
+        summary={summary}
+        tags={tags}
+        />
+      </div>
+    </div>
+  )
+}
+
+export default ExhibitGroup
