+++ src/css/Exhibit.css
@@ -1,0 +1,10 @@
+.exhibit-card {
+    height: 400px;
+    width: auto;
+}
+
+.exhibit-center {
+  margin-left: 50px;
+  margin-right: 50px;
+  padding: 50px;
+}
