--- src/css/ExhibitCard.css
@@ -1,4 +1,0 @@
-.exhibit-card {
-    height: 100px;
-    width: 100px;
-}
